class instance variables should be set on subclasses of paranoid classes

+++ lib/rails3_acts_as_paranoid.rb
@@ -14,8 +14,6 @@ module ActsAsParanoid
   def acts_as_paranoid(options = {})
     raise ArgumentError, "Hash expected, got #{options.class.name}" if not options.is_a?(Hash) and not options.empty?
     
-    return if paranoid?
-    
     class << self
       attr_accessor :paranoid_configuration, :paranoid_column_reference
     end
@@ -25,6 +23,8 @@ module ActsAsParanoid
     raise ArgumentError, "'time' or 'boolean' expected for :column_type option, got #{paranoid_configuration[:column_type]}" unless ['time', 'boolean'].include? paranoid_configuration[:column_type]
 
     @paranoid_column_reference = "#{self.table_name}.#{paranoid_configuration[:column]}"
+    
+    return if paranoid?
 
     ActiveRecord::Relation.class_eval do
       alias_method :delete_all!, :delete_all
@@ -88,14 +88,14 @@ module ActsAsParanoid
   module InstanceMethods
     
     def paranoid_value
-      self.send(self.class.base_class.paranoid_column)
+      self.send(self.class.paranoid_column)
     end
   
     def destroy!
       with_transaction_returning_status do
         run_callbacks :destroy do
-          self.class.base_class.delete_all!(:id => self.id)
-          self.paranoid_value = self.class.base_class.delete_now_value
+          self.class.delete_all!(:id => self.id)
+          self.paranoid_value = self.class.delete_now_value
           freeze
         end
       end
@@ -105,25 +105,25 @@ module ActsAsParanoid
       with_transaction_returning_status do
         run_callbacks :destroy do
           if paranoid_value.nil?
-            self.class.base_class.delete_all(:id => self.id)
+            self.class.delete_all(:id => self.id)
           else
-            self.class.base_class.delete_all!(:id => self.id)
+            self.class.delete_all!(:id => self.id)
           end
-          self.paranoid_value = self.class.base_class.delete_now_value
+          self.paranoid_value = self.class.delete_now_value
         end
       end
     end
     
     def recover(options={})
       options = {
-                  :recursive => self.class.base_class.paranoid_configuration[:recover_dependent_associations],
-                  :recovery_window => self.class.base_class.paranoid_configuration[:dependent_recovery_window]
+                  :recursive => self.class.paranoid_configuration[:recover_dependent_associations],
+                  :recovery_window => self.class.paranoid_configuration[:dependent_recovery_window]
                 }.merge(options)
 
       self.class.transaction do
         recover_dependent_associations(options[:recovery_window], options) if options[:recursive]
 
-        self.update_attributes(self.class.base_class.paranoid_column.to_sym => nil)
+        self.update_attributes(self.class.paranoid_column.to_sym => nil)
       end
     end
 
@@ -157,7 +157,7 @@ module ActsAsParanoid
     
   private
     def paranoid_value=(value)
-      self.send("#{self.class.base_class.paranoid_column}=", value)
+      self.send("#{self.class.paranoid_column}=", value)
     end
     
   end
